apps/db-tools/requirements.txt
fastapi
uvicorn[standard]
fastmcp
asyncpg
python-dotenv